fix: streaming settings (#123)
* Move streaming settings to account page

* Add warning for streaming

* Remove settings

* remove console log

---------
